Updated to the new Steam login requisites.
Previously just the basic form fields were mandatory, now it's required to
perform a successful login all the form fields in the POST request even the
empty ones.

I suspect that maybe not all fields are required and it's just
the `twofactorcode` field that is required and now data fields are
required to be oredered, the order goes as follows:

* `password`
* `username`
* `twofactorcode`
* `emailauth`
* `loginfriendlyname`
* `captchagid`
* `captcha_text`
* `emailsteamid`
* `rsatimestamp`
* `remember_login`

The `donotcache` field seems doesn't seem to be required and contains an
integer generated by `new Date.getTime()` function in JavaScript.