Outdated and depracated code needed to be updated.
Steam changed the group announcement API, they added more parameters in
the POST request for what it seems a feature that allows group owners
send their announcement in multiple languages.

Sourcemod depracated a cvar flag that was useless and was being
used because i didn't know it was useless also they changes the method
to reference functions as "variables" and added new functions to handle
this. There are still 2 warnings i need to address in the future since
they will be depracated.

Fixes #1