Adds the membership check module.
Adds 3 new functions to check if an account is present in a Steam group,
with these functions now the user has two ways to check, by reading
account profile or by reading the group profile, for the last option the
plugin stores the members list in a database by reading every page of the
XML generated by the steam api and parses them, to later be read without
much cpu affort.

Also adds an unfiniched function to retrieve the custom URL used in the
steam profile.